+++ database/migrations/2024_05_10_032256_add_is_import_column_to_purchase_requests_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('purchase_requests', function (Blueprint $table) {
-            $table->dropColumn('is_import');
-        });
+        if (Schema::hasTable('purchase_requests') && Schema::hasColumn('purchase_requests', 'is_import')) {
+            Schema::table('purchase_requests', function (Blueprint $table) {
+                $table->dropColumn('is_import');
+            });
+        }
     }
 };
