+++ database/migrations/2024_05_22_080120_add_description_column_to_permissions_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('permissions', function (Blueprint $table) {
-            $table->dropColumn('description');
-        });
+        if (Schema::hasTable('permissions') && Schema::hasColumn('permissions', 'description')) {
+            Schema::table('permissions', function (Blueprint $table) {
+                $table->dropColumn('description');
+            });
+        }
     }
 };
