+++ database/migrations/2024_07_30_012540_add_is_cancel_column_to_monthly_budget_reports_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('monthly_budget_reports', function (Blueprint $table) {
-            $table->dropColumn('is_cancel');
-        });
+        if (Schema::hasColumn('monthly_budget_reports', 'is_cancel')) {
+            Schema::table('monthly_budget_reports', function (Blueprint $table) {
+                $table->dropColumn('is_cancel');
+            });
+        }
     }
 };
