+++ database/migrations/2024_08_15_082731_add_dept_head_autograph_column_to_surat_perintah_kerja_komputer_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('surat_perintah_kerja_komputer', function (Blueprint $table) {
-            $table->dropColumn('dept_head_autograph');
-        });
+        if (Schema::hasColumn('surat_perintah_kerja_komputer', 'dept_head_autograph')) {
+            Schema::table('surat_perintah_kerja_komputer', function (Blueprint $table) {
+                $table->dropColumn('dept_head_autograph');
+            });
+        }
     }
 };
